+++ components/Icon.js
@@ -3,7 +3,7 @@ import * as Font from 'expo-font';
 import { createIconSetFromIcoMoon } from '@expo/vector-icons';
 import { Icon } from 'galio-framework';
 
-import GalioConfig from '../assets/fonts/galioExtra';
+import GalioConfig from '../assets/fonts/config/galioExtra.json';
 
 const GalioExtra = require('../assets/fonts/galioExtra.ttf');
 const IconGalioExtra = createIconSetFromIcoMoon(GalioConfig, 'GalioExtra');